pyproject.toml
[project]
name = "django-subscriptions-rt"
requires-python = ">=3.11"
keywords = []
license = {text = "MIT"}
readme = "README.md"
authors = [
    {name = "Reef Technologies", email = "opensource@reef.pl"},
]
classifiers = [
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dynamic = [
    "version",
]
dependencies = [
    "django>=3.2",
    "django-money>=3.0",
    "more-itertools>=10.6.0",
    "python-dateutil>=2.9.0.post0",
    "djangorestframework>=3.15.2",
    "requests>=2.32.3",
    "tenacity>=9.0.0",
    "pydantic>=2.10.6",
    "environs>=14.1.1",
]

[project.urls]
"Source" = "https://github.com/reef-technologies/django-subscriptions-rt"
"Issue Tracker" = "https://github.com/reef-technologies/django-subscriptions-rt/issues"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.version]
source = "scm"

[tool.pdm.dev-dependencies]
test = [
    "freezegun",
    "pytest",
    "pytest-apiver",
    "pytest-asyncio",
    "pytest-django",
    "pytest-xdist",
]
lint = [
    "codespell[toml]",
    "django-stubs>=1.12.0",
    "djangorestframework-stubs",
    "mypy>=1.8",
    "ruff",
    "types-freezegun",
    "types-python-dateutil",
    "types-requests",
]
release = [
    "towncrier",
]

[tool.pytest.ini_options]
pythonpath = ["."]  # allow for `import tests` in test files
target_package_name = "django_subscriptions_rt"  # required by pytest-apiver
DJANGO_SETTINGS_MODULE="tests.settings"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "D",
    "E", "F", "I", "UP",
    "TC005",
]
ignore = [
    "D100", "D105", "D107", "D200", "D202", "D203", "D205", "D212", "D400", "D401", "D415",
    "D101", "D102", "D103", "D104", # TODO remove once we have docstring for all public methods
    "D106",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"test/**" = ["D", "F403", "F405"]

[tool.codespell]
skip = "*.min.js,pdm.lock"
ignore-words-list = "datas"

[tool.towncrier]
directory = "changelog.d"
filename = "CHANGELOG.md"
start_string = "<!-- towncrier release notes start -->\n"
underlines = ["", "", ""]
title_format = "## [{version}](https://github.com/reef-technologies/django-subscriptions-rt/releases/tag/v{version}) - {project_date}"
issue_format = "[#{issue}](https://github.com/reef-technologies/django-subscriptions-rt/issues/{issue})"

[[tool.towncrier.type]]
directory = "removed"
name = "Removed"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Fixed"
showcontent = true

[[tool.towncrier.type]]
directory = "deprecated"
name = "Deprecated"
showcontent = true

[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "doc"
name = "Doc"
showcontent = true

[[tool.towncrier.type]]
directory = "infrastructure"
name = "Infrastructure"
showcontent = true

[tool.mypy]

[[tool.mypy.overrides]]
module = [
    "nox",
    "pytest",
    "tests.*",
]
ignore_missing_imports = true
